all tests passed

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: uyilmaz <uyilmaz@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/26 12:17:27 by uyilmaz           #+#    #+#             */
-/*   Updated: 2022/10/27 17:51:34 by uyilmaz          ###   ########.fr       */
+/*   Updated: 2022/10/31 13:09:22 by uyilmaz          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@ void	print_int(int n, int *count);
 void	print_char(char c, int *count);
 void	print_string(char *str, int *count);
 void	print_unsigned(unsigned int n, int *count);
-void	print_hex(size_t value, char c, int *count);
-void	print_ptr(size_t ptr, int *count);
+void	print_hex(unsigned long value, char c, int *count);
+void	print_ptr(unsigned long ptr, int *count);
 #endif
